toolboxfun/lib.py
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

def hello():
    '''function to test'''
    return 'Hello World'

def plot_history(history):
    plt.plot(np.sqrt(history.history['loss']))
    plt.plot(np.sqrt(history.history['val_loss']))
    plt.title('Model Loss')
    plt.ylabel('MSLE')
    plt.xlabel('Epoch')
    plt.legend(['Train', 'Val'], loc='best')
    plt.show()
